LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-subject-size-changes-expected.txt


FAIL View timeline with subject size change after the creation of the animation assert_equals: Effect at the midpoint of the active range expected "0.5" but got "0.3"
